feat(language-switcher): rename
