+++ packages/falcon-ecommerce-uikit/src/Footer/index.ts
@@ -2,4 +2,4 @@ export * from './Footer';
 export * from './Newsletter';
 export * from './FooterSections';
 export * from './Copyright';
-export * from './LanguageSwitcher';
+export * from './LocaleSwitcher';
